Move some interfaces around
